requirements.txt
playwright
requests
PyYAML
